move & improve script. add tests
