+++ macantine/management/commands/generate_geo_data_to_json.py
@@ -2,7 +2,7 @@ import json
 
 from django.core.management.base import BaseCommand
 
-from data.utils import clean_unicode_string
+from common.utils.utils import clean_unicode_string
 from common.api.datagouv import fetch_pats
 from common.api.decoupage_administratif import (
     fetch_communes,
